Bug 1233745 - Add a Travis config file & README badge
So we can check that `npm install` now succeeds under the new version of
nodejs, before we try it under the cedar-14 stack on Heroku.

It doesn't yet run `npm test` since the tests require actual
credentials, which we can either enter securely via Travis settings, or
else the tests could be rewritten to be entirely mocked.

Once this lands, someone with admin access to the autolander repo will
need to enable the Travis hook via:
https://travis-ci.org/profile