Bug 1237261 - Only process certain pull request events
When autolander receives a pull request event from GitHub, it may do one
of two things:
1) Comment on the PR if a bug number was not found
2) If a bug number was found, create an attachment on that bug if it
   does not already exist.

As such, only events that either create a new PR, or update the PR's
title or commit messages are relevant.

In addition, this helps avoid race conditions in cases where multiple
events are generated at the same time (for example when a PR has an
assignee set at the time of creation, both "opened" and "assigned"
events are sent).