Bug 1235619 - Only warn about missing bug numbers on PR creation
Previously if a PR was missing a bug number from it's title, upon every
PR change/event Autolander would spam the PR with a comment saying:

"""
Autolander could not find a bug number in your pull request title. All
pull requests should be in the format of: Bug [number] - [description].
"""

Rather than trying to scrape a PR's comments to see if a warning has
already been left, it's much easier to only leave a warning when the PR
is first created. It's highly unlikely that someone would add a bug
number initially and then remove it later.